Avoid extra allocations reading the cover image
Let's avoid allocation the `ext` variable; we can directly append it
to the `img_path` variable based on the MIME check.

Also avoid loading the image's size and data in memory if we are not
going to use them (i.e. if we are not actually the creating the image
file).