Pass an array of files rather than a single one to #get
Rather than passing a single file at a time and getting its tags
we could rather pass a whole array of files.

Accessing an add-on may be expensive since the function call is done
inside Node which goes to the C++ world and then the result is given
to the Node world.

Thus for N files, this is done N times. With this approach, this is
done only once.