Completely change the way #get works
Since the #get method now takes an array rather than a single file,
we can easily build a representation of a whole library easily.

This representation makes us save a lot of memory, avoiding duplicates
and avoids a lot of processing on the DaPlayer side since the data
is almost represented as wanted on the latter's side.