Allow the `get` method to take a progress callback
To give some feedback to the user, the `get` method can take an optional
callback that yields the current index and the total number of files.

Passing the total number of files may seem useless but this can be handy
if getting tags is handled in another process than where the files array
has been built.