Remove the `singles` key under artists
This array was containing indexes that were mapped back to the main
`singles` array but the problem is that if singles are removed, we
should either left an `undefined` value inside the array or change
every single index to decrement them.

This was here to simplify attaching singles to artists but this is
still easy anyway through `filter`.