Change the way we determine cover paths
Let's rather rely on the image's checksum as it is a built-in feature
of TagLib. Thus:

* Names are shorter.
* There's no longer any problem with artist or album names containing
  UTF-8 chars.
* There's no duplicated images inside the cache folder.

The only "regression" this brings is that it used to be possible to
associate images with albums without having to touch their tags but
anyway, we still have good performances as we cache the library on
disk inside DaPlayer.